Extract baseApplicationIcons into @tsmono/react/icons
The two app-level ApplicationIcons registries agreed on ~70 keys
verbatim; only fork and pendingTask had differing values. Move the
shared base (plus loggingIcons and iconForMimeType) into the package,
and let each app spread it and override or extend as needed. No
behavior changes for either app.
